#23 moved Mardao-Datastore PersistService and UnitOfWork class to GAE module.
