+++ guja-base/src/main/java/com/wadpam/guja/api/MonitorResource.java
@@ -23,10 +23,11 @@ package com.wadpam.guja.api;
  */
 
 
-import com.google.appengine.repackaged.org.joda.time.DateTime;
+
 import com.google.common.collect.ImmutableMap;
 import com.google.inject.Inject;
 import com.google.inject.name.Named;
+import org.joda.time.DateTime;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -37,7 +38,6 @@ import javax.ws.rs.Path;
 import javax.ws.rs.Produces;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
-import java.util.Date;
 
 
 /**
