+++ guja-be-ear/src/main/application/META-INF/application.xml
@@ -25,8 +25,8 @@
   <display-name>appengine-modules-ear</display-name>
   <module>
     <web>
-      <web-uri>guja-be-war-0.0.4-SNAPSHOT.war</web-uri>
-      <context-root>/guja-be-wacom.wadpam.gujar</context-root>
+      <web-uri>guja-be-war-0.0.6-SNAPSHOT.war</web-uri>
+      <context-root>/guja-be-wajar</context-root>
     </web>
   </module>
   <library-directory>lib</library-directory>
